fix: use unlock notify also on sqlite3_exec (#2021)
